1) Fixed timestamp name conflict on wifi probe, ScanResult timestamp renamed to tsf. 2) Changed Db Timstamp's column type from integer to float 3) Add schedule to ContactProbe 4) Telephone extends ImpluseProbe

+++ src/edu/mit/media/funf/storage/NameValueDatabaseHelper.java
@@ -54,7 +54,7 @@ public class NameValueDatabaseHelper extends SQLiteOpenHelper {
 			Arrays.asList(new Column(COLUMN_DATABASE_NAME, "TEXT"), // Name of this database
 						  new Column(COLUMN_INSTALLATION, "TEXT"), // Universally Unique Id for device installation 
 				      	  new Column(COLUMN_UUID, "TEXT"), // Universally Unique Id for file 
-					      new Column(COLUMN_CREATED, "INTEGER"))); // TIMESTAMP in data broadcast
+					      new Column(COLUMN_CREATED, "FLOAT"))); // TIMESTAMP in data broadcast
 	
 	
 	private final Context context;
